requirements.txt.txt
streamlit
ultralytics
opencv-python-headless
pillow
pandas
matplotlib